backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
PyPDF2==3.0.1
pandas==2.2.3
python-dotenv==1.0.1
httpx==0.27.2
ollama==0.3.3